README.md
I am still on the learning curve of data analysis, with a passion for health data, from curative to preventative. In order to make the best of evidenced-based care, we need to rely on real-time and accurate data. I look forward to getting any helpful insights, collaborations, and small tasks on the same.